Fix conditional frame removal logic
We want to remove empty frames in some cases.  Previously the logic
relied upon deleting entries from a table as we iterated over it.  This
caused items to be missed in certain cases.  Also, it was unnecessarily
complex.  Replace with a simple bool.

Also fixed some code which didn't get moved when it should have done.